+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and [A successful Git branching model](https://nvie.com/posts/a-successful-git-b
 
 ## [unreleased]
 
+## [0.9.3] - 2020-01-02
+### Added
+- Docker image (nix expression, build script, docker hub repo)
+
+### Fixed
+- Specified linux dependencies in nix files
+- Patch shebangs in Orthofinder scripts
+- Stack repl finds nix-provided zlib now
+- Venn Diagram list dependencies
+- Provide Fontconfig path to R scripts
+- Use shell.nix in test.sh
+- Marked a few tests as broken or nondeterministic
+
 ## [0.9.2] - 2019-12-10
 ### Changed
 - Friendlier install script